+++ AffiseAttributionLib/Classes/modules/advertising/AffiseAdvertisingApi.swift
@@ -4,4 +4,5 @@ import Foundation
 @objc
 public protocol AffiseAdvertisingApi: AffiseModuleApi {
     func startModule()
+    func getAdvertisingId() -> String?
 }