+++ AffiseAttributionLib/Classes/usecase/PackageInfoUseCase.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+internal protocol PackageInfoUseCase {
+    func getFirstInstallTime() -> TimeInterval?
+    func getPackageAppName() -> String?
+    func getAppVersion() -> String?
+    func getAppVersionRaw() -> String?
+}