+++ AffiseAttributionLib/Classes/usecase/PersistentUseCase.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+internal protocol PersistentUseCase {
+    func initialize(moduleManager: AffiseModuleManager)
+    func getAffDeviceId() -> String?
+}